+++ packages/nimble-components/src/table-column/menu-button/types.ts
@@ -1,0 +1,15 @@
+import type { MenuButtonToggleEventDetail } from '../../menu-button/types';
+
+/** @internal */
+export const menuSlotName = 'menu' as const;
+
+/**
+ * The type of the detail associated with the `menu-button-toggle` and `menu-button-beforetoggle`
+ * events on the menu button column.
+ */
+export interface MenuButtonColumnToggleEventDetail {
+    recordId: string;
+    newState: boolean;
+    oldState: boolean;
+    originalEvent: CustomEvent<MenuButtonToggleEventDetail>;
+}