+++ packages/nimble-components/src/table/components/row/index.ts
@@ -13,6 +13,8 @@ import { styles } from './styles';
 import { template } from './template';
 import type { TableCellState } from '../../../table-column/base/types';
 import type {
+    CellViewSlotRequestEventDetail,
+    RowSlotRequestEventDetail,
     TableActionMenuToggleEventDetail,
     TableFieldName,
     TableRecord,
@@ -263,6 +265,17 @@ export class TableRow<
             'transitionend',
             this.removeAnimatingClass
         );
+    }
+
+    public onCellViewSlotsRequest(column: TableColumn, event: CustomEvent<CellViewSlotRequestEventDetail>): void {
+        event.stopImmediatePropagation();
+        const eventDetails: RowSlotRequestEventDetail = {
+            // mkreis TODO: is this a valid non-null assertion?
+            rowId: this.recordId!,
+            columnInternalId: column.columnInternals.uniqueId,
+            slots: event.detail.slots
+        };
+        this.$emit('row-slots-request', eventDetails);
     }
 
     private readonly removeAnimatingClass = (): void => {