+++ packages/nimble-components/src/table/template.ts
@@ -12,6 +12,7 @@ import { tableHeaderTag } from './components/header';
 import { tableRowTag } from './components/row';
 import type { TableColumn } from '../table-column/base';
 import {
+    RowSlotRequestEventDetail,
     TableActionMenuToggleEventDetail,
     TableColumnSortDirection,
     TableRowSelectionMode,
@@ -172,6 +173,7 @@ export const template = html<Table>`
                                         @row-selection-toggle="${(x, c) => c.parent.onRowSelectionToggle(x.index, c.event as CustomEvent<TableRowSelectionToggleEventDetail>)}"
                                         @row-action-menu-beforetoggle="${(x, c) => c.parent.onRowActionMenuBeforeToggle(x.index, c.event as CustomEvent<TableActionMenuToggleEventDetail>)}"
                                         @row-action-menu-toggle="${(_, c) => c.parent.onRowActionMenuToggle(c.event as CustomEvent<TableActionMenuToggleEventDetail>)}"
+                                        @row-slots-request="${(_, c) => c.parent.onRowSlotsRequest(c.event as CustomEvent<RowSlotRequestEventDetail>)}"
                                         @row-expand-toggle="${(x, c) => c.parent.handleRowExpanded(x.index)}"
                                         :dataIndex="${x => x.index}"
                                     >
@@ -182,6 +184,12 @@ export const template = html<Table>`
                                                 slot="${x => `row-action-menu-${x}`}">
                                             </slot>
                                         `)}
+                                    `)}
+                                    ${repeat((x, c) => ((c.parent as Table).tableData[x.index]?.id ? (c.parent as Table).slotsByRow[(c.parent as Table).tableData[x.index]!.id] || [] : []), html<{ name: string, slot: string }>`
+                                        <slot
+                                            name="${x => x.name}"
+                                            slot="${x => x.slot}"
+                                        ></slot>
                                     `)}
                                     </${tableRowTag}>
                                 `)}