+++ packages/storybook/src/nimble/menu/menu.react.tsx
@@ -1,4 +1,4 @@
-import { Menu } from '@ni/nimble-components/dist/esm/menu';
+import { Menu } from '../../../../nimble-components/src/menu';
 import { wrap } from '../../utilities/react-wrapper';
 
 export const NimbleMenu = wrap(Menu);