+++ packages/storybook/src/nimble/table-column/anchor/table-column-anchor.stories.ts
@@ -1,12 +1,10 @@
 import { html, ref } from '@microsoft/fast-element';
 import type { HtmlRenderer, Meta, StoryObj } from '@storybook/html';
 import { withActions } from '@storybook/addon-actions/decorator';
-import { tableTag } from '@ni/nimble-components/dist/esm/table';
-import { tableColumnTextTag } from '@ni/nimble-components/dist/esm/table-column/text';
-import { AnchorAppearance } from '@ni/nimble-components/dist/esm/anchor/types';
-import { tableColumnAnchorTag } from '@ni/nimble-components/dist/esm/table-column/anchor';
-import { menuTag } from '@ni/nimble-components/dist/esm/menu';
-import { menuItemTag } from '@ni/nimble-components/dist/esm/menu-item';
+import { tableTag } from '../../../../../nimble-components/src/table';
+import { tableColumnTextTag } from '../../../../../nimble-components/src/table-column/text';
+import { AnchorAppearance } from '../../../../../nimble-components/src/anchor/types';
+import { tableColumnAnchorTag } from '../../../../../nimble-components/src/table-column/anchor';
 import {
     sharedTableActions,
     SharedTableArgs,
@@ -129,7 +127,7 @@ export const anchorColumn: StoryObj<AnchorColumnTableArgs> = {
             options: Object.keys(AnchorAppearance),
             control: { type: 'radio' },
             description:
-                'Set to `prominent` to make the anchor appear in a different color than normal text.'
+                'Set to `prominent` to make the anchor appear in a different color than normal text. This has no effect under the Color theme.'
         },
         underlineHidden: {
             name: 'underline-hidden',