+++ packages/storybook/src/nimble/table-column/menu-button/table-column-menu-button.react.tsx
@@ -1,0 +1,4 @@
+import { TableColumnMenuButton } from '../../../../../nimble-components/src/table-column/menu-button';
+import { wrap } from '../../../utilities/react-wrapper';
+
+export const NimbleTableColumnMenuButton = wrap(TableColumnMenuButton);