+++ packages/storybook/src/nimble/tooltip/tooltip.react.tsx
@@ -1,4 +1,4 @@
-import { Tooltip } from '@ni/nimble-components/dist/esm/tooltip';
+import { Tooltip } from '../../../../nimble-components/src/tooltip';
 import { wrap } from '../../utilities/react-wrapper';
 
 export const NimbleTooltip = wrap(Tooltip);