+++ packages/storybook/src/spright/rectangle/rectangle.stories.ts
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/html';
 import { html } from '@microsoft/fast-element';
-import { rectangleTag } from '@ni/spright-components/dist/esm/rectangle';
+import { rectangleTag } from '../../../../spright-components/src/rectangle';
 import { createUserSelectedThemeStory } from '../../utilities/storybook';
 
 interface RectangleArgs {