+++ packages/angular-workspace/nimble-angular/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @ni/nimble-angular
 
-This log was last generated on Tue, 21 May 2024 22:53:18 GMT and should not be manually modified.
+This log was last generated on Wed, 22 May 2024 18:32:15 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 24.3.2
+
+Wed, 22 May 2024 18:32:15 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v29.1.5
 
 ## 24.3.1